+++ bakedanuki/bakedanuki-util/python/bd_util/_test/test.py
@@ -11,19 +11,19 @@ def main():
     nodes = bdu.Nodes()
 
     trsf = nodes.create.transform(name="test")
-    values = [i for i in range(10000)]
+    keys = [(i, i) for i in range(10000)]
 
-    trsf.tx.keyframe.set_keys(values, frames=values)
-    trsf.ty.keyframe.set_keys(values, frames=values)
-    trsf.tz.keyframe.set_keys(values, frames=values)
+    trsf.tx.keyframe.set_keys(keys)
+    trsf.ty.keyframe.set_keys(keys)
+    trsf.tz.keyframe.set_keys(keys)
 
-    trsf.rx.keyframe.set_keys(values, frames=values)
-    trsf.ry.keyframe.set_keys(values, frames=values)
-    trsf.rz.keyframe.set_keys(values, frames=values)
+    trsf.rx.keyframe.set_keys(keys)
+    trsf.ry.keyframe.set_keys(keys)
+    trsf.rz.keyframe.set_keys(keys)
 
-    trsf.sx.keyframe.set_keys(values, frames=values)
-    trsf.sy.keyframe.set_keys(values, frames=values)
-    trsf.sz.keyframe.set_keys(values, frames=values)
+    trsf.sx.keyframe.set_keys(keys)
+    trsf.sy.keyframe.set_keys(keys)
+    trsf.sz.keyframe.set_keys(keys)
 
     nodes.modifier_manager.do_it_dag()
     nodes.modifier_manager.do_it_dg()